[contrib:http] using alias_method instead of prepend to patch
